+++ src/core/interfaces/JuegoInterfazMovimientos.py
@@ -3,9 +3,10 @@ from abc import ABC, abstractmethod
 class IJuegoInterfazMovimientos(ABC):
 
     @abstractmethod
-    def mover_ficha(self):
+    def realizar_movimiento(self):
         pass
 
     @abstractmethod
     def puede_hacer_algun_movimiento(self):
         pass
+